refactor: cleaner package structure
