+++ transport/kafka/producer_fn.go
