+++ transport/nats/publisher_fn.go
