+++ transport/nats/subscriber_fn.go
